add feat: add semantic voxel update, but not test

+++ nvblox/include/nvblox/core/voxels.h
@@ -57,7 +57,7 @@ static constexpr size_t kTotalNumberOfLabels = 21;
 typedef Eigen::Matrix<float, kTotalNumberOfLabels, 1> SemanticProbabilities;
 struct SemanticVoxel {
   // Initialize voxel to unknown label.
-  uint8_t semantic_label = 0u;
+  uint16_t semantic_label = 0u;
   // Initialize voxel to uniform probability.
   // Use log odds! So uniform ditribution of 1/kTotalNumberOfLabels,
   // should be std::log(1/kTotalNumberOfLabels)
